sim: deterministic replay pipeline + float64 env + sim eval fixes
Env (Tsimulation/pushshapes/env.py):
- float64 throughout: action_space, obs, step input, _get_obs
- set_state: angle before position (CoG rotation fix)
- get_episode_init(): full init state + obstacles + reset_seed + config_hash

Collection (Tsimulation/collect/):
- zarr_writer: float64 storage for state/actions/goal/reward
- zarr_writer: episode_init JSON attr for deterministic replay
- mouse_collect: pre-step obs, 2x window scale, float64 actions
- scripted_collect: float64 action computation, init_state metadata

Replay (Tsimulation/examples/replay_zarr.py):
- Use episode_init (float64 + reset_seed) when available
- total_frames trim for zero-padding
- --all mode with coverage + drift reporting

Sim eval (egomimic/eval/eval_sim.py):
- float64 for state init and actions
- Fixed reset seed for deterministic replay mode
- Filter batches to target embodiment only
- max_videos support
- pushshapes_sim_stick formatter
- eval_hnet_sim_cotrain.yaml: dual sim eval (circle + stick)

Tests updated: float64 dtype assertions.

Verified: 5/5 scripted + 11/11 mouse episodes replay with zero drift.

+++ Tsimulation/tests/test_smoke.py
@@ -75,11 +75,11 @@ def test_env_step_smoke(object_shape, pusher_shape, obstacle_level):
         obs, info = env.reset(seed=42)
 
         assert obs["agent_pos"].shape == (2,)
-        assert obs["agent_pos"].dtype == np.float32
+        assert obs["agent_pos"].dtype == np.float64
         assert obs["object_pose"].shape == (3,)
-        assert obs["object_pose"].dtype == np.float32
+        assert obs["object_pose"].dtype == np.float64
         assert obs["goal_pose"].shape == (3,)
-        assert obs["goal_pose"].dtype == np.float32
+        assert obs["goal_pose"].dtype == np.float64
         assert obs["image"].shape == (96, 96, 3)
         assert obs["image"].dtype == np.uint8
         assert "coverage" in info
